RLIB-19: Make struct rlib opaque and export only API calls
struct rlib is an opaque structure in the public header.
Only a few other structures are left in the public header
to support the documented API function: rlib_add_function().

On top of that, the Windows build also gets a DEF file to
ensure the order of exported symbols for a stable ABI.

rlib_mysql_report() and rlib_postgres_report() were removed as an
"easy" API and moved them to the only user, the Python binding.
They were not even documented as part of the public API.

A few structures and functions from pcode.[ch] are made part of
the public C API (but not other bindings) for the sole reason
to support the existing code in the PHP binding. It may be
revisited later.
