+++ libsrc/environment.c
@@ -1,5 +1,5 @@
 /*
- *  Copyright (C) 2003-2006 SICOM Systems, INC.
+ *  Copyright (C) 2003-2016 SICOM Systems, INC.
  *
  *  Authors: Bob Doan <bdoan@sicompos.com>
  *
@@ -17,12 +17,12 @@
  * Free Software Foundation, Inc., 59 Temple Place - Suite 330,
  * Boston, MA 02111-1307, USA.
  */
+#include <config.h>
 
 #include <unistd.h>
 #include <glib.h>
 
-#include "config.h"
-#include "rlib.h"
+#include "rlib-internal.h"
 
 static gchar * rlib_c_resolve_memory_variable(gchar *name) {
 	return (gchar *)g_getenv(name);
@@ -43,4 +43,8 @@ void rlib_new_c_environment(rlib *r) {
 	ENVIRONMENT(r)->rlib_resolve_memory_variable = rlib_c_resolve_memory_variable;
 	ENVIRONMENT(r)->rlib_write_output = rlib_c_write_output;
 	ENVIRONMENT(r)->free = rlib_c_free;
+}
+
+DLL_EXPORT_SYM struct environment_filter *rlib_get_environment(rlib *r) {
+	return ENVIRONMENT(r);
 }
