+++ libsrc/formatstring.c
@@ -34,7 +34,7 @@
 #include <monetary.h>
 #endif
 
-#include "rlib.h"
+#include "rlib-internal.h"
 #include "pcode.h"
 #include "rlib_langinfo.h"
