+++ libsrc/html.c
@@ -28,7 +28,7 @@
 #include <math.h>
 
 #include "config.h"
-#include "rlib.h"
+#include "rlib-internal.h"
 #include "rlib_gd.h"
 
 #define TEXT 1
