+++ libsrc/layout.c
@@ -35,7 +35,7 @@
 #include "config.h"
 
 #include "config.h"
-#include "rlib.h"
+#include "rlib-internal.h"
 #include "pcode.h"
 #include "rlib_input.h"
 #include "rlib_langinfo.h"
