+++ libsrc/libr.def
@@ -1,0 +1,56 @@
+LIBRARY "libr.dll"
+
+EXPORTS
+rlib_init@1
+rlib_init_with_environment@2
+rlib_get_environment@3
+rlib_free@4
+rlib_version@5
+rlib_bindtextdomain@6
+rlib_add_datasource_mysql@7
+rlib_add_datasource_mysql_from_group@8
+rlib_add_datasource_postgres@9
+rlib_add_datasource_odbc@10
+rlib_add_datasource_xml@11
+rlib_add_datasource_csv@12
+rlib_add_query_as@13
+rlib_add_query_pointer_as@14
+rlib_add_search_path@15
+rlib_add_report@16
+rlib_add_report_from_buffer@17
+rlib_set_radix_character@18
+rlib_set_output_format@19
+rlib_set_output_format_from_text@19
+rlib_set_output_parameter@20
+rlib_set_locale@21
+rlib_set_output_encoding@22
+rlib_set_datasource_encoding@23
+rlib_execute@24
+rlib_get_content_type_as_text@25
+rlib_spool@26
+rlib_get_output@27
+rlib_get_output_length@28
+rlib_add_function@29
+rlib_signal_connect@30
+rlib_signal_connect_string@31
+rlib_add_parameter@32
+rlib_add_resultset_follower_n_to_1@33
+rlib_add_resultset_follower@34
+rlib_query_refresh@35
+rlib_graph_add_bg_region@36
+rlib_graph_clear_bg_region@37
+rlib_graph_set_x_minor_tick@38
+rlib_graph_set_x_minor_tick_by_location@39
+rlib_setmessagewriter@40
+rlib_value_stack_pop@41
+rlib_value_stack_push@42
+rlib_value_new_number@43
+rlib_value_new_string@44
+rlib_value_new_date@45
+rlib_value_new_error@46
+rlib_value_free@47
+rlib_infix_to_pcode@48
+rlib_execute_pcode@49
+rlib_pcode_free@50
+rlib_graph@51
+rlib_chart@52
