+++ libsrc/navigation.c
@@ -21,7 +21,7 @@
 #include <string.h>
 
 #include "config.h"
-#include "rlib.h"
+#include "rlib-internal.h"
 #include "pcode.h"
 #include "rlib_input.h"
