+++ libsrc/resolution.c
@@ -22,7 +22,7 @@
 #include <ctype.h>
 
 #include "config.h"
-#include "rlib.h"
+#include "rlib-internal.h"
 #include "pcode.h"
 #include "rlib_input.h"
