+++ libsrc/value.c
@@ -26,9 +26,9 @@
 #include <string.h>
 #include <value.h>
 
+#include "rlib-internal.h"
 #include "config.h"
 #include "util.h"
-#include "rlib.h"
 
 
 /*--- string functions ---*/
