+++ src/examples/c/example.c
@@ -17,7 +17,8 @@
  * Free Software Foundation, Inc., 59 Temple Place - Suite 330,
  * Boston, MA 02111-1307, USA.
  */
- 
+
+#include <stdio.h>
 #include <rlib.h>
 #include <rlib_input.h>
